src/main/java/com/ctlms/DBConnCTLMS.java
package com.ctlms;

import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.SQLException;

public class DBConnCTLMS {
    
    private String jdbcURL = System.getenv("DB_URL");
    private String jdbcUsername = System.getenv("DB_USERNAME");
    private String jdbcPassword = System.getenv("DB_PASSWORD");

    public DBConnCTLMS(String jdbcURL, String jdbcUsername, String jdbcPassword) {
        this.jdbcURL = jdbcURL;
        this.jdbcUsername = jdbcUsername;
        this.jdbcPassword = jdbcPassword;
    }

    public Connection getConnection() throws SQLException, ClassNotFoundException {
        Connection connection = null;
        Class.forName("com.mysql.cj.jdbc.Driver");
        connection = DriverManager.getConnection(jdbcURL, jdbcUsername, jdbcPassword);
        return connection;
    }
}
